requirements.txt
asgiref==3.11.0
Django==6.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.4
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
uritemplate==4.2.0
python-decouple==3.8  # Environment variable management
dj-database-url==2.1.0  # Database URL parsing
